+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.2.6 Patch Notes
+## 1.3.0 Patch Notes
 - VOL-T
   - Added Crate Prefabs
   - Added an indicator to Static Snare
@@ -21,7 +21,9 @@
   - Made Signal Overload Hit Count scale with Attack Speed
   - Made Hit Impact VFX larger
   - Fixed Tempest Sphere Tick Rate not scaling with Attack Speed
-  - Fixed many bugs with Static Snare with high Attack Speed and or Cooldown Reduction
+  - Fixed Static Snare throw bug with high Attack Speed
+  - Fixed Static Snare throw bug with high Cooldown Reduction
+  - Fixed Static Snare assign bug with high Cooldown Reduction
   - Fixed Static Snare spamming NREs while dead
   - Fixed Signal Overload constructor error
   - Fixed Signal Overload doing more or less damage than expected
@@ -38,8 +40,10 @@
   - Berfed Sidestep duration from 0.15s to 0.1s
   - Nuffed Direct Current Projectile Size from 0.5m to 0.25m
   - Nerfed Max Charge capacity from 10 to 20 (Everything that scaled with Charge is scaled appropriately to be the same at max, this is a nerf because you are now encouraged to hit direct hits with Direct Current for 150% Charge gain compared to before, otherwise it's halved)
-  - Fixed some NREs
+  - Fixed Overdrive Exit throwing an NRE while dead
+  - Fixed being in heat throwing an NRE while dead <img src="https://i.postimg.cc/vmM6hpzc/image.png" width="20" height="20"/>
 - Items/Equipment
+  - Added Preserved Atoll
   - Added ProcSolver Support
   - Added a lip to the leftward border of Makeshift Plate
   - Buffed Amber Knife Chance from 10% to 15%